transport: split encoder into own module
